+++ frontend/src/components/image_upload/image_upload_container.js
@@ -1,11 +1,11 @@
-import { connect } from 'react-redux';
-import { uploadImage } from '../../actions/image_actions';
-import ImageUpload from './image_upload';
+// import { connect } from 'react-redux';
+// import { uploadImage } from '../../actions/image_actions';
+// import ImageUpload from './image_upload';
 
-const mapDispatchToProps = dispatch => {
-    return {
-        uploadImage: (image) => dispatch(uploadImage(image))
-    }
-};
+// const mapDispatchToProps = dispatch => {
+//     return {
+//         uploadImage: (image) => dispatch(uploadImage(image))
+//     }
+// };
 
-export default connect(null, mapDispatchToProps)(ImageUpload)
+// export default connect(null, mapDispatchToProps)(ImageUpload)
