+++ frontend/src/components/notes/note_compose.js
@@ -7,6 +7,7 @@ import ImageUploadContainer from '../image_upload/image_upload_container';
 class NoteCompose extends React.Component {
     constructor(props) {
         super(props);
+        this.fileInput = React.createRef();
 
         this.state = {
             title: '',
@@ -14,9 +15,14 @@ class NoteCompose extends React.Component {
             newNote: '',
             boardId: '',
             image: {},
+            file: null,
+            imageUrl: null,
+            errors: [],
+            inputReset: Date.now()
         };
 
         this.handleSubmit = this.handleSubmit.bind(this);
+        this.handleFiles = this.handleFiles.bind(this);
     }
 
     componentWillReceiveProps(nextProps) {
@@ -37,8 +43,42 @@ class NoteCompose extends React.Component {
         };
 
         this.props.makeNote(note)
+        .then((note) => {
+          debugger
+          if (this.fileInput.current.files.length <= [0]) {
+            const errors = [];
+            errors.push("Unable to upload image. image must be a JPEG or PNG and cannot be empty");
+            this.setState({ errors });
+
+          } else if (this.fileInput.current.files.length > 0) {
+
+            if (this.fileInput.current.files[0].type === 'image/jpg' ||
+              this.fileInput.current.files[0].type === 'image/png' ||
+              this.fileInput.current.files[0].type === 'image/jpeg') {
+              const image = new FormData();
+              image.append('image', this.state.file);
+              this.props.uploadImage(image, note.note.data).then(res => console.log(res))
+              // this.setState({
+              //   errors: [],
+              //   inputReset: Date.now(),
+              //   file: null,
+              //   imageUrl: null
+              // })
+              // this.setState({ file: e.target.files[0]})
+            } else {
+              const errors = []
+              errors.push('Invalid Image');
+              this.setState({ errors });
+            }
+          }
+        })
         // .then( () => this.props.closeModal())
 
+    }
+
+    handleFiles(e) {
+      e.preventDefault();
+      this.setState({ file: e.target.files[0] });
     }
 
     update(field) {
@@ -57,12 +97,18 @@ class NoteCompose extends React.Component {
                   value={this.state.title}
                   onChange={this.update('title')}
                   placeholder="Title of note"
-
                 />
                 <br />
                 <Image />
                 <br />
-                <ImageUploadContainer update={this.update('image')} />
+                <input type='file'
+                  ref={this.fileInput}
+                  key={this.state.inputReset}
+                  type='file'
+                  name='image'
+                  onChange={this.handleFiles}
+                />
+                {/* <ImageUploadContainer update={this.update('image')} /> */}
                 <br />
                 <input
                   type="textarea"
