+++ frontend/src/reducers/image_errors_reducer.js
@@ -3,7 +3,7 @@ import { RECEIVE_IMAGE_ERRORS } from '../actions/image_actions';
 const ImageErrorsReducer = (state = [], action) => {
     Object.freeze(state);
 
-    switch (action.key) {
+    switch (action.type) {
         case RECEIVE_IMAGE_ERRORS:
             return action.errors;     
         default:
