+++ frontend/src/util/note_api_util.js
@@ -17,6 +17,7 @@ export const editNote = data => {
 }
 
 export const updateNoteImage = (imageData, note) => {
+    debugger
     return axios.patch(`api/notes/${note._id}`, imageData)
 }
